Fixes per reviewer feedback
