+++ styles/_modal.scss
@@ -22,7 +22,7 @@
 
 .modal-site-name-section {
     background: #C3DACD;
-    border-top: 5px solid #388557;
+    border-top: 5px solid $green-light;
 }
 
 .modal-site-address-text {
@@ -31,7 +31,7 @@
     font-weight: normal;
     font-size: 16px;
     margin-bottom: 0;
-    color: #252525;
+    color: $black;
 }
 
 .modal-book-appointment-section {
